order-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.27.0
google-generativeai
openai
python-dotenv
python-jose[cryptography]
